refactor: used of shared component for status route and change class bages
